Move Memory store to a different folder
