+++ src/Neo.CLI/Settings.cs
@@ -11,7 +11,7 @@
 
 using Microsoft.Extensions.Configuration;
 using Neo.Network.P2P;
-using Neo.Persistence;
+using Neo.Persistence.Providers;
 using System;
 using System.Reflection;
 using System.Threading;
