+++ src/Neo/NeoSystem.cs
@@ -16,6 +16,7 @@ using Neo.Ledger;
 using Neo.Network.P2P;
 using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
+using Neo.Persistence.Providers;
 using Neo.Plugins;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
