+++ src/Neo/Persistence/Providers/MemoryStore.cs
@@ -18,7 +18,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
 
-namespace Neo.Persistence
+namespace Neo.Persistence.Providers
 {
     /// <summary>
     /// An in-memory <see cref="IStore"/> implementation that uses ConcurrentDictionary as the underlying storage.
