+++ src/Plugins/MPTTrie/Cryptography/MPTTrie/Trie.Proof.cs
@@ -10,7 +10,7 @@
 // modifications are permitted.
 
 using Neo.Extensions;
-using Neo.Persistence;
+using Neo.Persistence.Providers;
 using System;
 using System.Collections.Generic;
