+++ tests/Neo.Plugins.OracleService.Tests/TestBlockchain.cs
@@ -14,6 +14,7 @@ using Neo.Extensions;
 using Neo.Ledger;
 using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
+using Neo.Persistence.Providers;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
 using Neo.VM;
