+++ tests/Neo.Plugins.RpcServer.Tests/TestMemoryStoreProvider.cs
@@ -10,6 +10,7 @@
 // modifications are permitted.
 
 using Neo.Persistence;
+using Neo.Persistence.Providers;
 
 namespace Neo.Plugins.RpcServer.Tests
 {
